requirements.txt
python-telegram-bot==20.7
Flask==2.3.3
Flask-SQLAlchemy==3.1.1 